fix for mitaka unit test

+++ run_test.sh
@@ -3,24 +3,31 @@
 export CINDER_DIR=./cinder
 export CINDER_REPO_URL="https://git.openstack.org/openstack/cinder"
 
-echo "${CINDER_DIR}"
-if [ -d "${CINDER_DIR}" ]; then
-    rm -rf $CINDER_DIR
+echo "Running Flake8..."
+flake8 infortrend/
+if [ $? -ne 0 ]; then
+    exit 1
 fi
-
-git clone -b stable/mitaka $CINDER_REPO_URL --depth=1
+flake8 test/infortrend/
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+echo "Complete."
+   
+if [ -z "${2}" ]; then
+    if [ -d "${CINDER_DIR}" ]; then
+        echo "Deleting $CINDER_DIR"
+        rm -rf $CINDER_DIR
+    fi
+    git clone -b driverfixes/mitaka $CINDER_REPO_URL --depth=1
+else
+    echo "Skip Cloning cinder."
+fi
 
 source setupIFTdriver.sh $CINDER_DIR
 
 cd $CINDER_DIR
 
-if [ $(grep "'{posargs}' --concurrency" -q "tox.ini") ]; then
-    echo "Concurrency already set."
-else
-    # concurrency set to 4 for travisCI
-    sed -i "s#stestr run '{posargs}'#stestr run '{posargs}' --concurrency=4#g" tox.ini
-fi
-
-tox -e ${1} test_infortrend_
+tox -e ${1} -- -n cinder.tests.unit.volume.drivers.infortrend.test_infortrend_common --concurrency=4
 
 cd ..
